src/python/requirements.txt
# coap
aiocoap==0.3
cbor2==4.0.0
LinkHeader==0.4.3

# deployment
packaging==16.8

# env
PyYAML==3.12

# postgreSQL
aiopg==0.13.0
SQLAlchemy==1.1.10
psycopg2==2.7.1

# rest
aiohttp==2.1.0
cchardet==2.1.0
httpie==0.9.9
PyJWT==1.5.0
cryptography==1.9

# run
python-daemon==2.1.2
